fix bost tab - fetxh plugins
