+++ rubix-ui/frontend/src/components/hosts/host/flow/plugins/views/table.tsx
@@ -11,54 +11,58 @@ import { main, model } from "../../../../../../../wailsjs/go/models";
 import { PLUGIN_HEADERS } from "../../../../../../constants/headers";
 import RbTable from "../../../../../../common/rb-table";
 import { CreateModal } from "./create";
+import { useParams } from "react-router-dom";
+
+import PluginConf = model.PluginConf;
+import PluginUUIDs = main.PluginUUIDs;
 
 export const FlowPluginsTable = (props: any) => {
-  const { data, isFetching, connUUID, hostUUID, fetchPlugins } = props;
-  const [plugins, setPlugins] = useState([] as Array<model.PluginConf>);
-  const [pluginName, setPluginName] = useState<string>();
-  const [pluginsUUIDs, setPluginsUUIDs] = useState(
-    [] as Array<main.PluginUUIDs>
+  const { data } = props;
+
+  const { connUUID = "", hostUUID = "" } = useParams();
+  const [plugins, setPlugins] = useState([] as Array<PluginConf>);
+  const [selectedPlugins, setSelectedPlugins] = useState(
+    [] as Array<PluginConf>
   );
+  const [pluginName, setPluginName] = useState<string>();
+  const [pluginsUUIDs, setPluginsUUIDs] = useState([] as Array<PluginUUIDs>);
   const [networkSchema, setNetworkSchema] = useState({});
   const [isModalVisible, setIsModalVisible] = useState(false);
   const [isLoadingForm, setIsLoadingForm] = useState(false);
+  const [isFetching, setIsFetching] = useState(false);
 
-  let factory = new FlowPluginFactory();
-  let flowNetworkFactory = new FlowNetworkFactory();
+  const flowNetworkFactory = new FlowNetworkFactory();
+  const factory = new FlowPluginFactory();
+  factory.connectionUUID = connUUID;
+  factory.hostUUID = hostUUID;
 
   const columns = PLUGIN_HEADERS;
 
-  useEffect(() => {
-    if (isModalVisible) {
-      getSchema();
-    }
-  }, [pluginsUUIDs, isModalVisible]);
+  const rowSelection = {
+    onChange: (selectedRowKeys: any, selectedRows: Array<model.PluginConf>) => {
+      setPluginsUUIDs(selectedRows);
+      setSelectedPlugins(selectedRows);
+    },
+  };
+
+  const showModal = () => {
+    setIsModalVisible(true);
+  };
 
   const enable = async () => {
-    factory.connectionUUID = connUUID;
-    factory.hostUUID = hostUUID;
     await factory.BulkEnable(pluginsUUIDs);
     fetchPlugins();
   };
 
   const disable = async () => {
-    factory.connectionUUID = connUUID;
-    factory.hostUUID = hostUUID;
     await factory.BulkDisable(pluginsUUIDs);
     fetchPlugins();
   };
 
-  const rowSelection = {
-    onChange: (selectedRowKeys: any, selectedRows: Array<model.PluginConf>) => {
-      setPluginsUUIDs(selectedRows);
-      setPlugins(selectedRows);
-    },
-  };
-
   const getSchema = async () => {
     setIsLoadingForm(true);
-    if (plugins.length > 0) {
-      let plg = plugins.at(0) as unknown as model.PluginConf;
+    if (selectedPlugins.length > 0) {
+      let plg = selectedPlugins.at(0) as unknown as PluginConf;
       setPluginName(plg.name);
       const res = await flowNetworkFactory.Schema(connUUID, hostUUID, plg.name);
       const jsonSchema = {
@@ -69,9 +73,27 @@ export const FlowPluginsTable = (props: any) => {
     }
   };
 
-  const showModal = () => {
-    setIsModalVisible(true);
+  const fetchPlugins = async () => {
+    try {
+      setIsFetching(true);
+      let res = (await factory.GetAll()) || [];
+      setPlugins(res);
+    } catch (error) {
+      console.log(error);
+    } finally {
+      setIsFetching(false);
+    }
   };
+
+  useEffect(() => {
+    fetchPlugins();
+  }, []);
+
+  useEffect(() => {
+    if (isModalVisible) {
+      getSchema();
+    }
+  }, [pluginsUUIDs, isModalVisible]);
 
   return (
     <>
@@ -99,7 +121,7 @@ export const FlowPluginsTable = (props: any) => {
       <RbTable
         rowKey="uuid"
         rowSelection={rowSelection}
-        dataSource={data}
+        dataSource={plugins}
         columns={columns}
         loading={{ indicator: <Spin />, spinning: isFetching }}
       />
